day 9 not done yet

+++ src/Utils.kt
@@ -6,7 +6,7 @@ import java.io.File
 fun readInput(name: String) = File("src", "$name.txt").readLines()
 
 fun intcodeInput(input: List<String>) = intcodeInput(input.single())
-fun intcodeInput(input: String) = input.split(',').map { it.toInt() }
+fun intcodeInput(input: String) = input.split(',').map { it.toLong() }
 
 inline fun <reified T> finalAnswerIsNotWrong(answer: T, wrongAnswers: List<T>): T =
     answer
